requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
pydantic-settings==2.8.0
pyepp==0.1.6
xsdata==25.4
uvicorn==0.34.3
trio
asgi-lifespan
pytest-order